Windows backend works, Mac (and maybe Linux?) backend now broken

+++ test/test_monitor.rb
@@ -23,16 +23,27 @@ class TestMonitor < Test::Unit::TestCase
     should "should report a deletion" do
         changed_paths = []
         stopped = false
+        watcher = nil
         EM.run {
-            EMDirWatcher.watch TEST_DIR, ['/bar'], ['*.html'] do |changed_path|
+            watcher = EMDirWatcher.watch TEST_DIR, ['/bar'], ['*.html'] do |changed_path|
                 changed_paths << changed_path
                 EM.add_timer 0.2 do EM.stop end
             end
-            EM.add_timer 0.2 do
-                FileUtils.rm_rf File.join(TEST_DIR, 'bar')
+            timer_handler = lambda do
+                if not watcher.ready_to_use?
+                    EM.add_timer 0.2, &timer_handler
+                    return
+                end
+                EM.add_timer 0.5 do
+                  FileUtils.rm_rf File.join(TEST_DIR, 'bar')
+                  EM.add_timer 0.2, &timer_handler
+                  EM.add_timer 1 do EM.stop end
+                end
             end
-            EM.add_timer 1 do EM.stop end
+            EM.add_timer 0.2, &timer_handler
+            EM.add_timer 20 do EM.stop end
         }
+        watcher.stop
         assert_equal join(['bar/foo', 'bar/biz', 'bar/boo/bizzz'].sort), join(changed_paths.sort)
     end
