+++ test/test_tree.rb
@@ -239,7 +239,7 @@ class TestTreeRefreshing < Test::Unit::TestCase
 
   should "avoid traversing excluded directories" do
     @tree = EMDirWatcher::Tree.new TEST_DIR, nil, ['death']
-    FileUtils.ln_s(TEST_DIR, File.join(TEST_DIR, 'death'))
+    FileUtils.ln_s(TEST_DIR, File.join(TEST_DIR, 'death')) unless EMDirWatcher::PLATFORM == 'Windows'
     FileUtils.rm_rf File.join(TEST_DIR, 'bar', 'foo')
     changed_paths = @tree.refresh!
     assert_equal "bar/foo", join(changed_paths)
@@ -267,18 +267,20 @@ class TestTreeScopedRefresh < Test::Unit::TestCase
     @list = ['aa', 'biz', 'zz', 'bar/foo', 'bar/biz', 'bar/biz.html', 'bar/boo/bizzz'].sort
   end
 
-  should "fail with an exception when faced with an endless symlink loop" do
-    assert_raises Errno::ELOOP do
-      @tree = EMDirWatcher::Tree.new TEST_DIR
-      FileUtils.ln_s(TEST_DIR, File.join(TEST_DIR, 'bar', 'death'))
-      FileUtils.rm_rf File.join(TEST_DIR, 'bar', 'foo')
-      changed_paths = @tree.refresh! File.join(TEST_DIR, 'bar')
+  unless EMDirWatcher::PLATFORM == 'Windows'
+    should "fail with an exception when faced with an endless symlink loop" do
+      assert_raises Errno::ELOOP do
+        @tree = EMDirWatcher::Tree.new TEST_DIR
+        FileUtils.ln_s(TEST_DIR, File.join(TEST_DIR, 'bar', 'death'))
+        FileUtils.rm_rf File.join(TEST_DIR, 'bar', 'foo')
+        changed_paths = @tree.refresh! File.join(TEST_DIR, 'bar')
+      end
     end
   end
 
   should "report file deletion in inner directory when the scope specifies the directory" do
     @tree = EMDirWatcher::Tree.new TEST_DIR
-    FileUtils.ln_s(TEST_DIR, File.join(TEST_DIR, 'death'))
+    FileUtils.ln_s(TEST_DIR, File.join(TEST_DIR, 'death')) unless EMDirWatcher::PLATFORM == 'Windows'
     FileUtils.rm_rf File.join(TEST_DIR, 'bar', 'foo')
     changed_paths = @tree.refresh! File.join(TEST_DIR, 'bar')
     assert_equal "bar/foo", join(changed_paths)
@@ -286,7 +288,7 @@ class TestTreeScopedRefresh < Test::Unit::TestCase
 
   should "report file deletion in inner directory when the scope specifies the file" do
     @tree = EMDirWatcher::Tree.new TEST_DIR
-    FileUtils.ln_s(TEST_DIR, File.join(TEST_DIR, 'death'))
+    FileUtils.ln_s(TEST_DIR, File.join(TEST_DIR, 'death')) unless EMDirWatcher::PLATFORM == 'Windows'
     FileUtils.rm_rf File.join(TEST_DIR, 'bar', 'foo')
     changed_paths = @tree.refresh! File.join(TEST_DIR, 'bar', 'foo')
     assert_equal "bar/foo", join(changed_paths)
@@ -294,7 +296,7 @@ class TestTreeScopedRefresh < Test::Unit::TestCase
 
   should "not refresh the whole directory when the scope specifies a single file" do
     @tree = EMDirWatcher::Tree.new TEST_DIR
-    FileUtils.ln_s(TEST_DIR, File.join(TEST_DIR, 'bar', 'death'))
+    FileUtils.ln_s(TEST_DIR, File.join(TEST_DIR, 'bar', 'death')) unless EMDirWatcher::PLATFORM == 'Windows'
     FileUtils.rm_rf File.join(TEST_DIR, 'bar', 'foo')
     changed_paths = @tree.refresh! File.join(TEST_DIR, 'bar', 'foo')
     assert_equal "bar/foo", join(changed_paths)
@@ -302,7 +304,7 @@ class TestTreeScopedRefresh < Test::Unit::TestCase
 
   should "report file deletion in a subtree when the scope specifies a directory" do
     @tree = EMDirWatcher::Tree.new TEST_DIR
-    FileUtils.ln_s(TEST_DIR, File.join(TEST_DIR, 'death'))
+    FileUtils.ln_s(TEST_DIR, File.join(TEST_DIR, 'death')) unless EMDirWatcher::PLATFORM == 'Windows'
     FileUtils.rm_rf File.join(TEST_DIR, 'bar', 'boo', 'bizzz')
     changed_paths = @tree.refresh! File.join(TEST_DIR, 'bar')
     assert_equal "bar/boo/bizzz", join(changed_paths)
@@ -321,41 +323,43 @@ class TestTreeScopedRefresh < Test::Unit::TestCase
 
 end
 
-class TestTreeSymlinkHandling < Test::Unit::TestCase
+unless EMDirWatcher::PLATFORM == 'Windows'
+  class TestTreeSymlinkHandling < Test::Unit::TestCase
 
-  def setup
-    FileUtils.rm_rf TEST_DIR
-    FileUtils.rm_rf ALT_TEST_DIR
-    FileUtils.mkdir_p TEST_DIR
+    def setup
+      FileUtils.rm_rf TEST_DIR
+      FileUtils.rm_rf ALT_TEST_DIR
+      FileUtils.mkdir_p TEST_DIR
 
-    FileUtils.mkdir File.join(TEST_DIR, 'bar')
-    FileUtils.mkdir File.join(TEST_DIR, 'bar', 'boo')
+      FileUtils.mkdir File.join(TEST_DIR, 'bar')
+      FileUtils.mkdir File.join(TEST_DIR, 'bar', 'boo')
 
-    FileUtils.touch File.join(TEST_DIR, 'aa')
-    FileUtils.touch File.join(TEST_DIR, 'biz')
-    FileUtils.touch File.join(TEST_DIR, 'zz')
-    FileUtils.touch File.join(TEST_DIR, 'bar', 'foo')
-    FileUtils.touch File.join(TEST_DIR, 'bar', 'biz')
-    FileUtils.touch File.join(TEST_DIR, 'bar', 'biz.html')
-    FileUtils.touch File.join(TEST_DIR, 'bar', 'boo', 'bizzz')
+      FileUtils.touch File.join(TEST_DIR, 'aa')
+      FileUtils.touch File.join(TEST_DIR, 'biz')
+      FileUtils.touch File.join(TEST_DIR, 'zz')
+      FileUtils.touch File.join(TEST_DIR, 'bar', 'foo')
+      FileUtils.touch File.join(TEST_DIR, 'bar', 'biz')
+      FileUtils.touch File.join(TEST_DIR, 'bar', 'biz.html')
+      FileUtils.touch File.join(TEST_DIR, 'bar', 'boo', 'bizzz')
 
-    FileUtils.ln_s TEST_DIR, ALT_TEST_DIR
+      FileUtils.ln_s TEST_DIR, ALT_TEST_DIR
 
-    @list = ['aa', 'biz', 'zz', 'bar/foo', 'bar/biz', 'bar/biz.html', 'bar/boo/bizzz'].sort
+      @list = ['aa', 'biz', 'zz', 'bar/foo', 'bar/biz', 'bar/biz.html', 'bar/boo/bizzz'].sort
+    end
+
+    should "handle referencing root scope via symlink" do
+      @tree = EMDirWatcher::Tree.new TEST_DIR
+      FileUtils.rm_rf File.join(TEST_DIR, 'bar', 'foo')
+      changed_paths = @tree.refresh! ALT_TEST_DIR
+      assert_equal "bar/foo", join(changed_paths)
+    end
+
+    should "handle referencing root scope by real path when monitoring a symlinked path" do
+      @tree = EMDirWatcher::Tree.new ALT_TEST_DIR
+      FileUtils.rm_rf File.join(ALT_TEST_DIR, 'bar', 'foo')
+      changed_paths = @tree.refresh! TEST_DIR
+      assert_equal "bar/foo", join(changed_paths)
+    end
+
   end
-
-  should "handle referencing root scope via symlink" do
-    @tree = EMDirWatcher::Tree.new TEST_DIR
-    FileUtils.rm_rf File.join(TEST_DIR, 'bar', 'foo')
-    changed_paths = @tree.refresh! ALT_TEST_DIR
-    assert_equal "bar/foo", join(changed_paths)
-  end
-
-  should "handle referencing root scope by real path when monitoring a symlinked path" do
-    @tree = EMDirWatcher::Tree.new ALT_TEST_DIR
-    FileUtils.rm_rf File.join(ALT_TEST_DIR, 'bar', 'foo')
-    changed_paths = @tree.refresh! TEST_DIR
-    assert_equal "bar/foo", join(changed_paths)
-  end
-
 end
